Fix TitleProperties title type
